[cms] Add All Contractor Vote for contentious DCCs (#1104)
This adds the final portion laid out in the stakeholder approved proposal:
https://proposals.decred.org/proposals/fa38a3593d9a3f6cb2478a24c25114f5097c572f6dadf24c78bb521ed10992a4

When a DCC proposal for either issuance or revocation is deemed "contentious,"
administrators may now start a "Debate." This will then allow all currently active and valid
contractors to submit a vote to determine the final outcome of any given DCC proposal.

When the status for a DCC is updated to "debated" a current snapshot of the user base
is taken and the "weights" of each user's vote is determined. Currently, the weight is
simply the amount of billed hours over the previous 6 months of invoices that have been paid out.
We can use this raw amount however we like to best tabulate the results in a future PR.

As is the case with the previously implemented DCC functionality, administrators
currently have final say to Approve, Reject and start a Debate on a DCC. This
will be replaced with an automated system that will complete these without centralized
admin intervention. We'd first like to iron out the process without the worry of that automation.
Initial rigging for DebateDCC admin request