politeiawww: Fix route rexeg conflicts.
The proposal route regex was recently updated to allow variable length
tokens. The regex included all A-z characters. This caused requests to
other routes, such as "/proposals/tokeninventory" to be interpreted as a
proposal details request. This diff updates the token regex to only
allow hex.  The hex requirement combined with the minimum token length
requirement of 7 characters ensures that there will be no other
conflicts in the future.