[WIP] politeiawww: Update new user workflow.
This diff is a work in progress for an updated new user workflow. The
existing workflow allows an attacker to ascertain whether an email
address has been registered to politeia. It doesn't allow linking an
email address to a specific account, but it would still be better this
was not possible at all. The new user workflow is also unnecessarily
complex, which has lead to user identity bugs in the past.