gitbe: Limit git fsck during anchoring.
This diff prevents the git fsck command from being run during every
hourly anchoring. It is instead run once per day during the 06:58 UTC
anchoring. This time was chosen because it occurs in the middle of the
night for the USA and Brazil, which is likely the majority of politeia
traffic.