gitbe: Fix UnvettedExists bug.
This diff fixes the UnvettedExists gitbe method. It was previously
checking the unvetted repo for the record on the master branch. This is
incorrect. Unvetted records are stored in the unvetted repo as branches.
Once they become vetted records then they are merged into master. The
UnvettedExists method now checks the unvetted repo branches for the
record.

It also updates a few plugin calls, which were calling the
UnvettedExists method when they should have been calling the VettedExits
method.