politeiad: Update cache build process.
When the cache needs to be rebuilt it would use the backend `Inventory`
call to retrieve the full record inventory, including all versions of
every record, and store them in memory. This was a temporary solution
that has now started to become a problem due to the amount of data that
exists on the mainnet proposals site.

This diff adds two calls, `UnvettedTokens` and `VettedTokens` to the
backend interface that return the tokens of all unvetted and vetted
records, respectively. When the cache gets rebuilt it will use these
calls to fetch the record tokens then retrieve a single record at a time
to insert into the cache. The decred plugin cache rebuild process was
also updated to do the same thing. It retrieves the decred plugin data
and builds the decred plugin cache for each record individually.