[cms] Fix batch proposal passthrough (#1242)
This PR fixes the batch proposal passthrough's lack of CSRF for the POST request.

With new added passthrough handler, both GET and POST requests to proposals 
are dealt with as needed.  GET requests are passed straight through, whereas POST
first require requesting a GET /version to receive the required cookies and CSRF. 
Those then get added to the POST request and is passed along.  The raw response
from proposals is then passed back up and out.