[cms] Remove cache reference for tlog (#1282)
This is the first diff for clearing the way for tlog implementation for cms.

First, we are now rebuilding the cmsdb with direct records from the backend. 
Previously we were getting the records from the cache, but now it's direct 
from the files themselves.

Next, we are now just requesting invoice information from the CMS database 
where all invoice information was stored already.