[cms] Add Codetracker infrastructure (#1185)
This PR adds the ability to track development statistics of all CMS users that
have their domain set to "Developer". There is currently and updater/crawler
that runs upon every startup and at the start of every month at 12:01 AM GMT.
The crawler updates the codetracker statistics based on config settings for
organizations and repositories. Currently the codetracker is implemented for
the Github API. If we ever need to move to a new code repository solution,
it would just be a matter of adding a new implementation of the CodeTracker.

Back on the cmsww side, code stats are saved in the cms userdb. These
stats are updated once the crawler has finished processing and updating the
CodeTracker databse (that stores the raw PR, review and commit information).

Administrator users are able to request a range of dates of code stats per
user. This will be used for review developer invoices to ensure their progress
matches with hours billed.