politeiawww: Remove CSRF on public routes.
Previously, all routes were CSRF protected. This is not necessary and is
actually detrimental to privacy since it allows proposal votes being
trickled in through tor to be grouped by the CSRF token that they were
cast with.

This diff removes CSRF protection from public routes while keeping it
for all authenticated routes.