politeiawww: Memoize token inventory.
This diff memoizes the token inventory call in politeiawww. I'm not a
fan of this change since it requires invalidating the memoized results
all over the place which is bound to lead to bugs down the road, but
this only a temporary change until tlog is launched to help deal with
the decrediton 1.6 token inventory polling, so it should be fine for
now.