tstore: Delete vetted record content bug.
When a record is made public the encrypted blobs in the kv store are
re-saved as clear text, but the tlog leaf remains the same since the
record content did not actually change. The backend was only deleting
one of these blobs when a vetted record was being censored. This diff
updates the tstore to delete both the encrypted blob and clear text
blob.