ticketvote: Fix active votes cache bug.
This diff removes all panics from the active votes cache and replaces
them with graceful failures.

When a ballot is cast the record is locked by the tstore backend, but
this does not apply to the active votes cache which is lazily updated on
read requests. This can lead to a vote being removed from the active
votes cache while another thread is still validating a ballot for the
vote, causing the panics to be hit.

This diff is pretty ugly. These concurrency issues will be resolved and
cleaned up for good once a plugin command and all cache updates are
executed using a single kv store transaction.