multi: Add receipt to vote details.
This diff adds a receipt to the StartDetails structure that is saved to
disk when a vote is started.

The start vote command is the only plugin command that doesn't save a
receipt to prove the command payload was accepted by the server. This
was how it was previously implemented in the git backend and was carried
over to the tstore backend implementation. I don't know why it was
previously implemented like this, it could have been an oversight or it
could be because its an admin command, but I think it makes sense to add
it for the sake of consistency and because the start vote command may
not always be an admin command (ex. informal stakeholder polls on
pi-reddit).