multi: Add timestamp to ticket vote.
This diff adds a unix timestamp to the cast vote details that is saved
to disk when a ticket vote is cast.

The timestamp was not included in the git backend ticket votes because
the git backend only flushed the vote journals once per hour. This means
that you could only determine if a vote was cast within a +/- 60 minute
period. Adding the timestamp to the vote provided granular details that
could be used to correlate blocks of votes to their owner.

This argument does not apply to the tstore backend because tlog already
provides a granular timestamp for each vote leaf. Since the data is
already public it should be returned through the API so that clients
like dcrdata, who use the data to construct vote graphs, do not need to
walk the tlog tree manually in order to parse the timestamp.