politeiawww: Fix legacy pagination bug.
Multiple legacy routes were previously returning 500s because the legacy
API does not paginate properly in a lot of routes, but the politeiad v2
API expects the requests to be paginated, causing errors to occur.

This diff updates the function that is used to request proposals from
politeiad for the legacy routes to paginate the requests properly. It
also updates the `pictl legacytest` command that tests the legacy routes
to specifically test the pagination issue.