Update modules.
This diff updates the go version to 1.15, updates the dcrdata module to
v6, and runs a `go mod tidy`.

We ran into some weird module errors when trying to import the
politeiawww client into dcrdata. Turns out that one of the dcrdata
dependencies, `github.com/dgraph-io/badger`, doesn't follow semver and
it was causing all kinds of issues. What a pain! Upgrading to dcrdata v6
should fix the issue.