multi: Fix set status json typo.
This diff fixes a typo in in the JSON of the `Reason` field of a record
status change. This changes the structure that is saved to disk, but the
status change reason is only required when censoring or abandoning
proposals and no proposals have been censored or abandoned yet since the
launch of the new politeia site. We can still fix this before it becomes
a permanent bug.