politeiavoter: Update prompt pass placement.
This diff moves the dcrwallet password prompt to the begining of the
voting function so that the user does not have to wait for the expensive
O(n) wallet calls to complete in order to input their password. It also
adds in a dcrwallet call that verifies the password is correct.

This patch was submitted by moo31337.