pictl: Add save as csv option to vote results.
This diff adds a `--save` option to the `pictl voteresults` command that
instructs the command to save the vote results to disk as a csv.  A
`--filter` option was also added that can be used in conjunction with
`--save` if the user only wants to save the results of a specific vote
option.