politeiad: Setup fscks.
This diff adds a `--fsck` flag to politeiad that, when provided,
performs synchronous filesystem checks on startup for all record and
plugin data. This includes verifying the coherency of cached data.

The fscks have not been implemented in this diff, but notes have been
added to the various fsck functions to help with the implementation.