politeiad: Remove politeiaimport tool.
This diff removes the `politeiaimport` tool from the politeiad cmds.
This was a dev tool that was used to copy public git repos into a git
backend locally. The git backend is no longer supported so there is no
use for this tool anymore.

The politeiad `sharedconfig` package, which only contained two exported
variables, was also removed since the only reference to it was in the
`politeiaimport` tool.