politeiawww: Add page sizes to policy replies.
The page sizes for various routes in the `records`, `comments`, and
`ticketvote` APIs are hardcoded in the APIs as variables. This allows
golang clients to references the variables. 

This commit adds the page sizes to the policy replies so that the gui
can reference the policy response rather than hardcode the page sizes.

This diff adds a new policy route to the records API to return the API's
page sizes.  In adittion, this adds a new pictl command 
`pictl recordpolicy` which calls the new route mentioned above.