politeiawww: Add logger package.
This diff moves the logging state into into a new `logger` package
instead of storing it in the global main scope. This allows external
packages, such as plugins, to setup their own loggers without having to
add any code to main. This will allow plugins to be completely self
contained.