pi: Add API to retreive billing status changes.
Clients need to be able to retrieve the full billing status metadata so
that the admin username can be displayed and so that the admin signature
is available if the client wants it. In order to do this, API needs to
be added to the pi plugin for retrieving the full billing status change
metadata objects.  With that said, the current `StatusReason` in not
needed in the pi `Summary` and can be removed as this feature wasn't
deployed to production yet.

**Implementation:**

   - Remove `StatusReason` from pi API `Summary`.
   - Add new `/billingstatuschanges` route to the pi API.
   - Add new pictl command `pictl proposalbillingstatuschanges "token"`
     command which consumes the new pi API.