politeiawww: Add legacy package.
This diff moves all of the code for the legacy APIs into a `legacy`
package. No functionality or API was changed, the codebase was just
re-organized.  This allows us to start with a fresh politeiawww server
context as we re-write the user layer to use a plugin architecture and
to allow for horizontal scaling. Once we migrate to the new user layer
we can simply delete the `legacy` package and be done with it. No
untangling required.