pi: Allow for billing status error correction.
**Background:**

Let's say an admin makes a mistake and sets the billing status to
completed when they actually meant to set it to closed. There should be
a way for them to correct this mistake.

We would like to enable error correction by adding a plugin setting
which limits the number of the allowed billing status changes. It
defaults to 1 and can be increased by a sysadmin in rare cases where a
billing status error correction is needed.

**Implementation:**

 - Adds `allowedbillingstatuschanges` pi plugin setting  which defaults
   to 1 and can be increased if a billing status error correction is
   needed.

 - Adds in billing status validation that allows for the following state
   transitions:

    - active can only be set to closed or completed.
    - closed and completed can be set to any **other** billing status
      when the plugin setting allows for it.

 - Adds `active` status to the `pictl proposalsetbillingstatus` command.