politeiawww: Add websockets package.
This diff moves the websocket code out of main and into a `websockets`
package. A `Manager` context is used to manage the websocket connections
instead of of having them stored directly on the politeiawww context.

Packaging the websocket code like this will make it easier to move the
websocket connections into a data store that allows for horizontal
scaling of the politeiawww servers and will allow us to add a websocket
plugin framework at some point in the future.