multi: Move identity methods.
This diff moves two identity methods out of the `util` package and in
into more appropriate places.

`IdentityFromString` was moved into the `identity` package with the
other identity methods and was renamed to `PublicIdentityFromString` in
order to be consistent with the existing naming convetion.

`RemoteIdentity` was moved into `politeiawww/identity.go` since this is
the only place in the codebase that it was being invoked. It was renamed
to `remoteIdentity`.