logger: Fix nil pointer bug.
This fixes a nil pointer bug that was hit when a package attempts to
use its logger without having the log rotator setup.

Now that individual packages initialize the logger themselves instead of
having it initialized in main, it's possible for package loggers to be
used without ever having the log rotator setup. An example is when CLI
tools use politeiawww packages like when dbutil uses the userdb
implemenations.