Add number of bunches that vote.
This code flips differential time tickers to N go routines that expire
after some pre-determined time. We are moving to this model because
differential time tickers have proven brittle. When a differential
ticker does not fire everything else gets hung behind it but with this
model an individual hang will not impact subsequent votes.

The rough implementation is:
* Votes are equally divided into bunches
* Every bunch has a random start time and random duration
* Votes fire after bunch start + random duration
	- If the vote fails it is retried in the same loop shortly after
	  the failure
	- If the vote succeeds the go routine shuts down