politeiavote: Fix delay bug.
This updates the vote command to cast trickled votes in a separate go
routine.

The vote trickler does not account for the time it takes to actually
cast the vote, i.e. the time it takes to send the request to the server
and to receive a response, when it calculates the delays to insert in
between each vote. This adds in unexpected delay that can cause the
trickler to be unable to cast all votes before the voting period ends.

Casting the vote in a separate go routine fixes the issue.