politeiawww: Pull sessions store apart.
The previous sessions store was tangled up with the legacy user database
and politeia specific session values.

This moves the tangled session store implementation into the legacy
package and implements a generic session store that uses a generic DB
interface for encoded session values and does not make any assumptions
about the application specific session values. This implementation will
be used as we turn the user layer into plugins.

This is really how it should have been implemented from the start. I'm
the one that wrote the original implementation, so you know... what a
dummy.