pictl: Fix vote summary printing.
This commit fixes a bug that was introduced by 63d933ce that caused the
percent signs in the vote summary printing to be printed incorrectly.

`printf(voteSummaryString())`

The above code is same thing as:

`fmt.Printf("%\n") // Prints "%!(MISSING)"`

These should be:

`printf("%v\n", voteSummaryString())`

It also cleans up the `voteSummaryString()` function and replaces some
hardcoded arguments in the `rfptest` command with variables.