pi: Add approved proposal status.
This commit adds an 'approved' proposal status to the pi plugin.

Previously, the proposal status was considered to be 'active', meaning
the proposal is active and allowed to be billed against, once the
proposal vote was approved. This is fine for standard proposals. RFP
proposals, however, do not request funds and do not bill against the
treasury. The 'active' status does not apply to them. Once an RFP vote
is approved, the RFP proposal status is 'approved'. This does not apply
to RFP submissions sinces RFP submissions do bill against the treasury.