pictl: Add proposal status opt to seedproposals.
This commit adds a --proposalstatus option to the `pictl seedproposals`
command. If provided, all proposals submitted by the command will be set
to the specified proposal status. This makes it easier to fine tune your
testing environment, which is desirable when needing to test specific
features or bug fixes.

It also makes two additional changes to the `seedproposals` command:

- Updates the comment and comment like defaults to be more reasonable.

- Switches the hard coded proposal statuses out for the newly added pi
  plugin proposal statuses.