politeiad: Execute batch reads concurrently.
This commit updates the politeiad batch read route to execute the read
commands concurrently instead of serially in order to improve
performance.

Each batch of read commands is executed using a sync WaitGroup counter
to track the completion status of the commands and a shared memory cache
to aggreate the replies. The memory cache is local to each read batch.