Fix typo.
This commit updates the politeiad batch read route to execute the read
commands concurrently instead of serially in order to improve
performance.

A batch of read commands is executed using a sync WaitGroup counter to
track the completion status of the commands and a shared memory cache to
aggregate the replies. Each batch uses its own unique memory cache.

This performance optimization cut the execution of a 5 command read
batch by ~50%.