politeiavoter: Fix retry loop bug.
This commit fixes a bug that was preventing the failed vote retry loop
from exiting when a SIGINT (ctrl+c) is received. The user was being
forced to use a SIGKILL to kill the politeiavoter process.

The root cause was that the retry loop was using a time.Sleep() call
instead of using the WaitFor() call, which accepts a context and exits
when the context is canceled.