politeiavoter: Improve summary printing.
This commit updates the summary printing to consider already voted
errors as being successfully cast votes.

Already voted errors are not considered to be failures because they
occur when a network error or dropped client connection causes
politeiavoter to incorrectly think that the first attempt to cast the
vote failed. politeiavoter will attempt to retry the vote that it has
already successfully cast, resulting in the already voted error.

Printing these errors as failures is misleading to the user, who will
likely consider a failure as meaning the vote was not successfully cast,
which is incorrect in this case.