politeiavoter: Fix duplicate vote bug.
This commit fixes a bug that was causing politeiavoter to continually
retry casting a vote when a duplicate vote error is returned.

Duplicate vote errors can occur when the first attempt results in a
network error where the vote is actually cast, but politeiavoter reports
it as failed. Example, the client connection drops after the vote has
already been submitted to the backend. politeiavoter will attempt to
re-submit these votes and duplicate vote errors will be returned.

Previously, a duplicate vote error was considered a success case and the
vote was no longer attempted to be re-cast.

bc7da5b introduced a change that prevented the success case code path
from being reached on duplicate vote errors. This resulted in
politeiavoter continually retrying these votes when it should have
simply exited. This commit fixes this bug and reverts it to the previous
behavior.