pi: Cache proposal statuses.
Background:

This diff aims to improve a performance issue where the pi summary
command for proposals in the "Approved" politeiagui tab required two
different calls to be made to the tstore backend in order to determine
their status. The first call retrieves record metadata and the second
call retrieves billing status metadata. Each of these calls result in
the tlog tree being retrieved. A ticketvote vote summary request is
also needed to get the current vote status of the proposals.

Implementation:

This commit adds an in-memory cache which is used to cache proposal data
required to determine the proposal status at runtime such as record
metadata, vote metadata, the vote status and the proposal billing status
changes. The cache is necessary to improve the performance of
determining a status of a proposal at runtime by reducing the number of
expensive backend calls that result in the tlog tree be retrieved, which
gets very expensive when a tree contains tens of thousands of ticket
vote leaves. This is helpful when the cached data is not expected to
change, which means that once we store the data in cache we don't need
to fetch it again. The cache entries are lazy loaded and it's capacity
is limited to 1000 entries.

Benchmark:

We have tested this with 5 proposals that have an active proposal status
and with ~1000 tlog leaves each.  The pi summaries call took ~675ms
without the proposal status cache and ~200ms with the proposal status
cache.  A ~70% improvement.