politeiavoter: Improve hoursprior UX.
This commit improves the --hoursprior UX by improving the help and error
messages associated with the --hoursprior setting as well as adding
validation that prevents the user from specifying both an --hoursprior
and a --voteduration at the same time.

If a --voteduration is provided, the trickler will cast all votes within
that duration.

If a --voteduration is not provided, the trickler will cast all votes
within the remaining time minus the --hoursprior setting.

--hoursprior still defaults to 12 hours.

**Example 1**

A vote has 6 days remaining. The user attempts to trickle in their votes
over a 3h period. Previously, they would get the following error.

```
not enough time left to trickle votes: -9h0m0s < 12h0m0s, use
--hoursprior to modify this behavior
```

Using this commit, no error is returned and the votes are trickled in
over the 3h period.

**Example 2**

A vote has 6 hours remaining. The user attempts to trickle in their
votes without specifying a --voteduration or --hoursprior. Previously,
they would get the following error.

```
not enough time left to trickle votes: -6h0m0s < 12h0m0s, use
--hoursprior to modify this behavior
```

With this commit, they now get the following error message.

```
the --hoursprior setting (currently set to 12) specifies that all votes
should be cast within 12h0m0s prior to the end of the voting period. The
remaining voting period of 6h0m0s does not allow for this. Adjust
--hoursprior to modify this behavior.
```